% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{limno_palettes}
\alias{limno_palettes}
\title{Complete list of palettes}
\format{
An object of class \code{list} of length 9.
}
\usage{
limno_palettes
}
\description{
Use \code{\link[=limno_palette]{limno_palette()}} to construct palettes of desired length.
}
\keyword{datasets}
