% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palEx.R
\name{palEx}
\alias{palEx}
\title{Color Palettes extractor}
\usage{
palEx(file, n)
}
\arguments{
\item{file}{\code{.jpg} file of image}

\item{n}{Number of colors extracted}
}
\value{
A list of color values extracted from a user provided image (jpg format) based on k-means clustering of the image.
}
\description{
A function to extract a Color palette from a input \code{.jpg} file.
}
\seealso{
\link[stats:kmeans]{stats::kmeans}
}
