% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{print.palette}
\alias{print.palette}
\title{Print a palette}
\usage{
\method{print}{palette}(x, ...)
}
\arguments{
\item{x}{Desired palette}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print a palette
}
